+++ test-app/mocks/mockClients/dashboards/definitions/examples/bar/index.js
@@ -1,4 +1,5 @@
 // @ts-nocheck
 const { barChartDashboard } = require('./bar-chart-dashboard')
+const { barChartStackedDashboard } = require('./bar-chart-stacked-dashboard')
 
-module.exports = { barChartDashboard }
+module.exports = { barChartDashboard, barChartStackedDashboard }