+++ test-app/mocks/mockClients/dashboards/definitions/parent-child/child-dashboard-1.js
@@ -1,0 +1,36 @@
+// @ts-nocheck
+const { lists } = require('../examples/visualisations')
+
+const childDashboard1 = {
+  id: 'child-dashboard-one',
+  name: 'Child one dashboard',
+  description: 'Dashboard used for mocking child dashboards',
+  sections: [
+    {
+      id: 'child-one-section-1',
+      display: 'Child one - Section 1',
+      description: 'Child one - Section 1 description',
+      visualisations: [],
+    },
+    {
+      id: 'child-one-section-2',
+      display: 'Child one - Section 2',
+      description: 'Child one - Section 2 - description',
+      visualisations: [],
+    },
+    {
+      id: 'parent-section-1',
+      display: 'Child one - Section 2',
+      description: 'Child one - Section 2 - description',
+      visualisations: [],
+    },
+    {
+      id: 'all-data',
+      display: 'All Data',
+      description: '',
+      visualisations: [lists.fullDataset],
+    },
+  ],
+}
+
+module.exports = childDashboard1